result: amg with a function
This function run required that the agent run and do exactly what
the function said, including a decision and resources. This is a far
end of the spectrum in terms of control, because the agent has little
control. Next we are going to try an intermediate approach that returns
a retry or stop, but allows the agent to decision on resources.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>